Refactor latency queue headers
